requirements.txt
# VaaniSeva - Python Dependencies
# AI for Bharat Hackathon 2026

# ── Core (used by Lambda & seed script) ──
boto3>=1.34.0
botocore>=1.34.0
twilio>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0

# ── Dev / Testing ──
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
